fix: isolate hosted tool compatibility
- move hosted tool stubs into a focused FastAPI utility
- keep oversized endpoint and config tests line-neutral against main
- cover unsupported backend stubs, include stripping, and restore cleanup
